Replace find buildfiles implementation to use readdir().
The std::filesystem-provided directory iterator was noticeably slower
in going through the directories, so this implements it 'old school'.
Might need some ifdefs however in the future to gurantee portability
everywhere.

We still map that back to std::filesytem::path, so might be worthwhile
to investigate if that causes performance issues.

For this, first time using absl in this project, which means we also
need a compilation database for the language server to be happy. Added
that as well in this CL.