Parse 'in' and 'not in'.
For this, a parse conflict hat to be resolved in the for-loop.
Previously, the list of variables was just a list of expressions,
so the ambiguity `for a in ["x"]` could, with one lookahead, be
interpreted as the expression (a in ["x"]) or the variable
'a' bound in the for-loop over ["x"].
Solved by limiting the expression between for and in to be
a comma list (or a tuple of) identifiers.

Parse in and 'not in', the latter being a single token coming
from the scanner.