NamedLineIndexedContent: wrap of (filename, content, line-column-map)
Content that is processed inevitably needs to be able to print
locations for the user. For that, the filename and LineColumnMap
have been kept alongside historically to be able to have everything
together when it was necessary to print a location. Other than
that, the filename or the index would never be needed alone in itself.

So wrap them up together to better semantically reflect what this is
and is used for ... and also to make it much easier to print locations.