Don't std::move around strings anymore once there are string_view refs to it.
Subtle.
string_views in the AST refer to the original content to
be able to show locations. The content is owned by ParsedBuildFile that
was stored (and move'ed) by value in the ParsedProject map.

Typically, this is fine, as moving a std::string keeps its dynamically allocated
byffer at the same place when moving, so string_views pointing to it
point to the same place. But, there is also short string optimization...
In that case, the array with content is moving with the memory.

Conclusion: keep the ParsedBuildFile as unique_ptr<> in the container.