Use absl::Cleanup() for some C resources.
While at it, have CollectFilesRecursive() return the vector of
collected files and drop the count all walked files return value which
was an early form of collecting stats, but didn't really belong there.
Collect these now in the place that is interested in it: the caller's
predicates get called, so they can do that work.