Make dwyu testable and add test.
Break out internal functionality into a DWYUGenerator class to
make it testable into dwyu-internal.h. This then allows to mock
out reading source files that headers are extracted from.

Create a testutil for EditCallbacks so that we can easily test
expected edits.

Write a test-fixture in dwyu_test.cc to pull it all together and
include a simple test for adding and removing dependencies. To
be expanded.

Overall: no functional change, just make it testable.