Add a test for non-removal if unaccounted headers are there.
If there is a header, but globally there is no library that is actually
exporting it, hold off with removal suggestions as we can't be sure.
This should usually not happen, but maybe the header is generated
by a dependency (similar to cc proto lib) that we don't understand.

This was already handled correctly, this just adds a test for ist.