Allow multiple libraries to export the same header.
There _should_ be a 1:1 relationship of headers to libraries, but
sometimes there can be situation in which two libraries export the
same header (e.g. a production and debug version of a library).

Allow for this to not have a conceptually OneToOne but OneToN
relationship.

This affects the output of lib-headers, but also how dwyu deals
with these
  * If there are multiple alternatives and none of them are
    used in a dependency list, we can't auto-suggest something but
    need to let the user decide.
  * Or: if multiple libraries are given in a dependency list providing
    the same header, this is at least worth the attention of the user
    as we can't decide to remove anything.