Work around some absl/strings:string_view ambiguity.
In absl/strings:string_view, there is the string_view.h exported.

But it is _also_ exported by absl/strings:strings but with the remark
that this is only there for backward compatibility. In fact, it is
mentioned twice, in hdrs and in textual_hdrs.

We use this fact to make sure bant correctly identifies
absl/strings:string_view as only source for the dependency to add.