Handle use-case of deprecated alias pointing to new library.
Usually, if there is an alias that points to a library, it is considered
a perfectly valid alternative. However, if that is marked with a
deprecation notice, the pointee is the only useful alternative.

Make sure to emit edits to remove the deprecated and add the new
in that case.