Only complain about workspace external projects if issues with and without bzlmod
We look at both old (WORKSPACE) and new (WORKSPACE.bzlmod, MODULE.bazel).
The layout of the directories might be different, so if the project supports
both forms, one of them might not yield all external projects (depending on
how bazel was run).
Only complain if _all_ of them fail to yield something.