Centralize look-up of sourcce location in preparation for expression eval.
Before: each parsed build file was asked directly for look-up of substrings
within its content to provide a foo/BUILD:121:42 location.

In the future, there will be content that is not directly in the
file itself (e.g. globbed filenames), so a string_view found in
the AST of that file might not be in that file content range, but
stored somewhere else.

So: have parsed-project keep track of all the (disjoint) ranges of
strings that represent content, and map that to a SourceLocator (which
is essentially providing that line/col look-up). The SourceLocator then
is responsible to map any sub-string within that to a correct line/column
location.

Wire that up to the current ParsedProject pointing to each of the
ParsedBuildFile contents.

Make the ParsedBuildFile SourceLocator private and only visible to
the ParsedProject to not tempt callers to ask the file directly -- after
expression eval, any substring they encounter in the AST
might not actually not be in that build file, so asking
the ParsedProject is the only way to report a location.

No new functionality yet but
  * Add DisjointRangeMap that provide a mapping of string_views to
    a value (here: we use SourceLocator*) and provide a look-up
    mechanism to find by any subrange
  * Break out the SourceLocator functionality from NamedLineIndexContent
    and make a virtual base class, as soon we need a version for
    generated content.