Prepare extracting location information from binary operations.
To report a location, we need to have some string-view covering
the expression. For that, add a string_view for the range in BinOpNode.

This is filled in the parser, but faily simple right now: mostly
covered is the range the operation token itself represents (such as '+').
This is good enough for now, but should probably expanded to the whole
range covering the lhs and rhs of these expressions.