Implement string concatenation.
This creates a new string, so that is an entirely new sequence
of characters not found in the original file so mapping that
to a printable <file>:<line>:<col>: source location would not
work if we look in all our files.

So register a SourceLocator with the project to map this new
string_view to a FixedSourceLocator that reports the location
of the operator in the original file.

Allocate string and SourceLocator in our arena for compactness
and simple lifetime management.