When matching header files, try to find the longest suffix match.
Sometimes headers have a different prefix than expected, so they won't
match exactly with what we create in the lib header index.

Provide a way to just match a long enough suffix (should at least
have one / in it). For that, have an optional flag to reverse the
key in a lib header index creation for faster lookup.