Construct a minimal set of sufficient dependencies.
The headers in the sources are provided by libraries. Ideally, that is
only one, but unfortunately some projects are organized that there can
be multiple targets providing a header; we call these 'alternative'
libraries.

This was handled before, but this change makes more effort to
minimize the set of alternatives we consider: if header a is provided
by [target_x, target_y], and header b is provided by [target_x, target_z],
we just need [target_x] to satisfy needs for a and b.

In the course of that found that protocol buffers have a target that
provides the headers, but, since it is a header-only library, don't
actually provides a linkable dependency. Exclude this for now manually,
but there should be some better way to deal with it, maybe some sort
of viability-score for libraries.