Refine compilation DB: also emit virtual includes.
Virtual includes is a concept in which bazel create a different view
of an include path, so the compilation DB needs to point to that
symbolic link tree.

Check if --absl=1 is in .bazelrc and emit -DGTEST_HAS_ABSL=1

Some basic protobuf workaround, as we don't fully see through
proto libraries (but this needs to be solved more fundamentally).