compilation-db: resolve include paths from workspace directories.
In newer bazels with modules, the directory names are not exactly
the project names anymore (they include ~ and possibly version name),
so always use the directory names to assemble includes for
compilation DBs.

Also, upgrade the dependency graph to optionally call back with
any dependency it finds; use that in the compilation DB to selectively
only use includes provided by targets we acctually need.