Start supporting built-in macros for richer see-through of rules.
This allows to process some non-standard rules without having
to go through the trouble of fully seeing through *.bzl files.

Instead, provide a purely declarative way to provide enough information
for the rest of bant to be able to provide dwyu features.

For now, the built-in macros are in bant/builtin-macros.bnt which is
compiled into the binary.

  * Added a file bant/builtin-macros.bnt that is compiled into
    the binary for further parsing. Define a simple macro syntax
    that can be fully expressed with the existing syntax.
  * Have parsed-project parse that at start-up and provide a
    look-up macro-name -> Node*
  * Add a variable substitution that can take a tree and replace
    variables and return a new copy. This can be used to expand
    macros (but also later: ListComprehensions).
  * Add a macro substitutor that takes a BUILD-file ast, and
    replaces all rule-calls that are macros in it with the
    macro content; variables inside there are subsituted from the
    kwargs of the original call.
  * Each component tested, but short of wiring it up to
    the dwyu tooling ... this change already gets too big :)

So follow-up: wire up to dwyu, and fill in some macros. Possibly
getting rid of some of the previously necessary proto buffer
and grpc hacks.