Expand dependency rounds with io-threads if requested.
Opening files can take a while on network file systems, so following
hundreds of files can take a while. So allow to read them in threads.

For that, break up the operations: open files in threads, then
parse and elaborate in a single thread (parsing does not need io,
so is not hindered by a slow file system).