Replace glob(3) with CollectFilesRecursive() and GlobMatchBuilder
Use our own infrastructure to provide glob functionality. This allows
for easier transition to a filesystem abstraction later.

While at it, rename some things to better convey the meaning
`want_dir_p` -> `enter_dir_p` to better indicate that this is used
in the recursive walking context.

`want_file_p` became `want_file_or_dir_p` and gained the
functionality to also collect directories (needed by implementation
of Glob() that mimicks glob(3)).

Also rename `BuildDirectoryMatchPredicate()` to
`BuildRecurseDirMatchPredicate()` in the `GlobMatchBuilder` to better
reflect what it is doing.

In the process also made the selection criteria for a project directory
more strict in the workspace (exclude double-tilde directories). This
became apparent as the order files arrived is now dependent on the
random readdir() sequence not the (possibly sorted ?) glob() sequence.